requirements.txt
streamlit==1.47.1
moviepy==2.2.1
opencv-python==4.12.0.88
Pillow==11.3.0
python-pptx==1.0.2
requests==2.31.0